+++ package.json
@@ -13,6 +13,7 @@
   },
   "dependencies": {
     "@babel/compat-data": "^7.11.0",
+    "@bcrs-shared-components/corp-type-module": "1.0.3",
     "@mdi/font": "^5.5.55",
     "@sentry/browser": "^5.21.1",
     "@sentry/integrations": "^5.21.1",
