+++ src/enums/index.ts
@@ -6,4 +6,5 @@ export * from './roleTypes'
 export * from './incorporatorTypes'
 export * from './modes'
 export * from './nameRequestStates'
+export * from './nameRequestTypes'
 export * from './routeNames'
