+++ src/enums/nameRequestTypes.ts
@@ -1,0 +1,6 @@
+export enum NameRequestTypes {
+  BENEFIT_COMPANY = 'BC',
+  BC_COMPANY = 'CR',
+  BC_ULC_COMPANY = 'UL',
+  BC_CCC = 'CC',
+}
